+++ gatsby-config.js
@@ -41,6 +41,7 @@ module.exports = {
       resolve: `gatsby-transformer-remark`,
       options: {
         plugins: [
+          `gatsby-plugin-catch-links`,
           `gatsby-remark-smartypants`,
           `gatsby-remark-component`,
           `gatsby-remark-responsive-iframe`,
