+++ package.json
@@ -21,7 +21,7 @@
     "storybook:build": "build-storybook -c .storybook -o public/docs"
   },
   "dependencies": {
-    "@carbon/addons-website": "^0.0.1-alpha.13",
+    "@carbon/addons-website": "^0.0.1-alpha.14",
     "@carbon/colors": "0.0.1-alpha.31",
     "@carbon/grid": "0.0.1-alpha.31",
     "@carbon/icons-react": "0.0.1-alpha.31",
@@ -41,6 +41,7 @@
     "fs": "0.0.1-security",
     "gatsby": "2.0.70",
     "gatsby-image": "2.0.22",
+    "gatsby-plugin-catch-links": "^2.0.9",
     "gatsby-plugin-google-analytics": "^2.0.8",
     "gatsby-plugin-manifest": "^2.0.11",
     "gatsby-plugin-offline": "^2.0.18",
