--- src/components/Tile/index.js
@@ -1,3 +1,0 @@
-import Tile from './Tile';
-
-export default Tile;
