+++ src/components/WebsiteTile/index.js
@@ -1,0 +1,3 @@
+import WebsiteTile from './WebsiteTile';
+
+export default WebsiteTile;
