+++ src/components/WebsiteTile/website-tile.scss
@@ -39,7 +39,12 @@
   }
 }
 
-a.bx--tile {
+.tile a,
+a.tile {
+  text-decoration: none;
+}
+
+.bx--tile {
   display: flex;
   flex-direction: column;
   justify-content: space-between;
@@ -49,18 +54,27 @@ a.bx--tile {
   background: $ibm-colors__white;
   z-index: 1;
   padding: 0; //needs to stay zero for aspect ratio to be correct
+  transition: $transition--base;
+}
 
-  &:hover {
-    background: $ui-03;
-  }
+a:hover .bx--tile {
+  background: $ui-03;
 }
 
 .bx--tile .bx--aspect-ratio--object {
-  padding: $spacing-md;
+  padding: $spacing-md 30% $spacing-md $spacing-md;
+
+  @include breakpoint('lg') {
+    padding-right: $spacing-md;
+  }
+
+  @include breakpoint('xlg') {
+    padding-right: 30%;
+  }
 }
 
 .tile__title {
-  @include type-style('body-short-02');
+  @include type-style('heading-03');
   display: block;
   width: 100%;
   text-decoration: none;
@@ -69,9 +83,8 @@ a.bx--tile {
   padding: 0 9% 0 0;
 }
 
-.tile--feature .tile__title,
-.tile--feature--alt .tile__title {
-  @include type-style('heading-03');
+.tile__title--small {
+  @include type-style('body-short-02');
 }
 
 .tile__caption {
@@ -136,21 +149,21 @@ a.bx--tile {
 }
 
 // Dark
-a.bx--tile--clickable--dark {
+.bx--tile--dark {
   background: $ibm-colors__gray-90;
   color: $ibm-colors__gray-10;
 }
 
-a.bx--tile--clickable--dark:hover {
+a:hover .bx--tile--dark {
   background: $ibm-colors__gray-80;
 }
 
-a.bx--tile--clickable--dark .tile__title,
-a.bx--tile--clickable--dark .tile__caption {
+.bx--tile--dark .tile__title,
+.bx--tile--dark .tile__caption {
   color: $ibm-colors__gray-10;
 }
 
-a.bx--tile--clickable--dark .tile__link-icon svg {
+.bx--tile--dark .tile__link-icon svg {
   fill: $ibm-colors__white;
 }
 
@@ -190,7 +203,7 @@ a.bx--tile--clickable--dark .tile__link-icon svg {
     }
   }
 
-  a.bx--tile {
+  .bx--tile {
     @include breakpoint('md') {
       position: absolute;
       bottom: 0;
@@ -206,7 +219,7 @@ a.bx--tile--clickable--dark .tile__link-icon svg {
 
 .ibm--col-lg-8 {
   @include breakpoint('lg') {
-    a.bx--tile {
+    .bx--tile {
       width: 47.5%;
     }
   }
@@ -214,7 +227,7 @@ a.bx--tile--clickable--dark .tile__link-icon svg {
 
 .ibm--col-lg-12 {
   @include breakpoint('lg') {
-    a.bx--tile {
+    .bx--tile {
       width: 32%;
     }
   }
@@ -222,7 +235,7 @@ a.bx--tile--clickable--dark .tile__link-icon svg {
 
 .ibm--col-lg-16 {
   @include breakpoint('lg') {
-    a.bx--tile {
+    .bx--tile {
       width: calc(25% - 2rem);
     }
   }
@@ -245,17 +258,35 @@ a.bx--tile--clickable--dark .tile__link-icon svg {
 }
 
 .tile__feature-content {
-  padding: 1rem 1rem 4rem 1rem;
-
   @include breakpoint('md') {
+    padding: 1rem 12% 1rem 1rem;
     width: 50%;
-    padding-right: 12%;
-    padding-bottom: 1rem;
   }
 
   @include breakpoint('lg') {
-    padding-right: 25%;
+    padding-right: 12.5%;
   }
+}
+
+.tile--feature h2.bx--type-expressive-heading-05 {
+  padding: 1rem 1rem 0 1rem;
+  color: $text-01;
+  font-weight: 300;
+
+  @include breakpoint('md') {
+    padding: 0;
+  }
+
+  // Expressive paragraph 01 at tablet and below, responsive type includes are breaking
+  @media (max-width: 66rem) {
+    line-height: 125%;
+    letter-spacing: 0;
+    font-size: calc(1.5rem + 0.25 * ((100vw - 20rem) / 46));
+  }
+}
+
+.tile--feature .background--inverse h2.bx--type-expressive-heading-05 {
+  color: $ibm-colors__gray-10;
 }
 
 .tile__feature-img {
@@ -272,7 +303,7 @@ a.bx--tile--clickable--dark .tile__link-icon svg {
   margin-top: 0;
 }
 
-.tile--feature a.bx--tile {
+.tile--feature .bx--tile {
   position: absolute;
   bottom: 0;
   right: 0;
@@ -330,6 +361,11 @@ a.bx--tile--clickable--dark .tile__link-icon svg {
 
 .tile--feature--alt h2 {
   margin-bottom: 1.5rem;
+  color: $text-01;
+}
+
+.tile--feature--alt .background--inverse h2 {
+  color: $ibm-colors__gray-10;
 }
 
 .tile--feature--alt h2 strong {
@@ -364,7 +400,7 @@ a.bx--tile--clickable--dark .tile__link-icon svg {
   }
 }
 
-.tile--feature--alt a.bx--tile {
+.tile--feature--alt .bx--tile {
   position: absolute;
   bottom: 0;
   right: 0;
