+++ src/styles/index.scss
@@ -53,7 +53,7 @@ $font-family: 'IBM Plex Sans', 'Helvetica Neue', Arial, sans-serif;
 @import '../components/BackToTop/back-to-top.scss';
 @import '../components/TitleBlock/title-block.scss';
 @import '../components/Icon/icon.scss';
-@import '../components/Tile/tile.scss';
+@import '../components/WebsiteTile/website-tile.scss';
 @import '../components/SiteHeader/header.scss';
 
 //---------------------------------------
