+++ src/templates/page.js
@@ -8,7 +8,7 @@ import Snippet from '../components/CodeSnippet';
 import PageTable from '../components/PageTable';
 import BackToTop from '../components/BackToTop';
 import TitleBlock from '../components/TitleBlock';
-import Tile from '../components/Tile';
+import WebsiteTile from '../components/WebsiteTile';
 import AnchorLinks from '../components/AnchorLinks';
 import Icon from '../components/Icon';
 import { Grid, Column } from '../components/Grid/Grid';
@@ -41,7 +41,7 @@ const renderAst = new rehypeReact({
     'anchor-links': AnchorLinks,
     'title-block': TitleBlock,
     icon: Icon,
-    tile: Tile,
+    tile: WebsiteTile,
     grid: Grid,
     column: Column,
   },
